타일 사이즈수정

+++ shot/tile.h
@@ -8,13 +8,13 @@ class tile
 public:
 	Vector2 globalPos;
 	Vector2 localPos;
-	float height;
+	int height;
 	int tiletype;
 	//static Texture2D texture;
 	static void textureLoad();
 	tile();
-	tile(int type, Vector2 globalpos,float h);
+	tile(int type, Vector2 globalpos,int h);
 	void drawtile(float scale);
 	void update();
-	void init(int type, Vector2 globalpos, float h);
+	void init(int type, Vector2 globalpos, int h);
 };
